move direct (old) sampling functions to examples and remove tests for those functions in test file
